float

+++ player2.py
@@ -30,7 +30,7 @@ class GameConnection(Protocol):
 			os._exit(1)
 		if data.find("angle: ") != -1:
 			tmp_str = data.split(" ")
-			self.gs.ball.angle = int(tmp_str[1])
+			self.gs.ball.angle = float(tmp_str[1])
 			self.gs.ball.rotate()
